Final commit

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -26,19 +26,17 @@ public class AdminController {
         this.roleService = roleService;
     }
 
-    // Показать всех пользователей + форму для создания нового
     @GetMapping({"", "/"})
-    public String showAllUsers(Model model) {
+    public String showAdminPanel(Model model) {
         List<User> users = userService.getAllUsers();
         model.addAttribute("users", users);
+        User admin = userService.findByEmail("admin@mail.ru");
+        model.addAttribute("currentUser", admin);
         model.addAttribute("user", new User());
-        model.addAttribute("roles", roleService.getAllRoles());
         model.addAttribute("allRoles", roleService.getAllRoles());
         return "admin/admin";
     }
 
-
-    // Создать нового пользователя
     @PostMapping("/new-user")
     public String createUser(@ModelAttribute("user") User user,
                              @RequestParam(value = "rolesSelected", required = false) List<Long> rolesSelected) {
@@ -46,25 +44,23 @@ public class AdminController {
             Set<Role> roles = new HashSet<>(roleService.getRolesByIds(rolesSelected));
             user.setRoles(roles);
         } else {
-            user.setRoles(new HashSet<>()); // если роли не выбраны - пустой набор
+            user.setRoles(new HashSet<>());
         }
         userService.saveUser(user);
         return "redirect:/admin";
     }
 
-    // Форма редактирования пользователя
     @GetMapping("/update/{id}")
     public String editUserForm(@PathVariable("id") Long id, Model model) {
         User user = userService.getUserById(id);
         if (user == null) {
-            return "redirect:/admin"; // если пользователь не найден - редирект на список
+            return "redirect:/admin";
         }
         model.addAttribute("user", user);
         model.addAttribute("roles", roleService.getAllRoles());
         return "admin/update-user";
     }
 
-    // Обновить пользователя
     @PostMapping("/update")
     public String updateUser(@ModelAttribute("user") User user,
                              @RequestParam(value = "rolesSelected", required = false) List<Long> rolesSelected) {
@@ -78,10 +74,12 @@ public class AdminController {
         return "redirect:/admin";
     }
 
-    // Удалить пользователя
     @PostMapping("/delete/{id}")
     public String deleteUser(@PathVariable("id") Long id) {
-        userService.deleteUser(id);
+        User user = userService.getUserById(id);
+        if (user != null) {
+            userService.deleteUser(id);
+        }
         return "redirect:/admin";
     }
 }
